SCC-2189 Remove Avro base64 encoding
By default the avro client base64 encodes records passed to it, and this should be the default behavior. But when passing records to Kinesis, these are base64 encoded again. To avoid this, this update sets a flag to be sure that the avro encoded records are not base64 encoded before being passed to the kinesis client.