SCC-2189 Add basic holding parser
This implements a basic parser for the holdings records, mainly so that the output object can pass the Avro schema validation. It simply takes the data in the 863, 866 and legacy "h" fields and passes them into the object as strings.

This also will make it easier to find and display this data in the discovery API/SCC front-end.