NOREF Check for empty hash instead of nil
When checking for a missing `853` field the previous fix mistakenly evaluates the value for `nil` instead of checking for an empty hash. This is resolved by using `empty?` on the variable instead, which should match the return from the `filter` call.

This also removes a minor logging issue with a few values being written directly to STDOUT